requirements.txt
matplotlib==3.7.1
PyYAML==6.0.2
simpy==4.1.1
pytest==8.3.4